requirements.txt
google-genai
python-dotenv
numpy
pandas
python-docx
langchain
tiktoken
transitions
psycopg2-binary
sqlalchemy
pgvector